pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taskfile"
version = "0.3.91"
description = "Universal Taskfile runner with multi-environment deploy support. CI/CD agnostic — run locally or from any pipeline."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    {name = "Tom Sapletta", email = "tom@sapletta.com"},
]
keywords = ["taskfile", "deploy", "ci-cd", "devops", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "clickmd>=1.1.1",
    "fixop>=0.1.0",
]

[project.optional-dependencies]
ssh = [
    "paramiko>=3.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
]
llm = [
    "litellm>=1.30.0",
]
dev = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "build>=1.0",
    "twine>=5.0",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
    "pyqual>=0.1",
]

[project.scripts]
taskfile = "taskfile.cli:main"
tf = "taskfile.cli:main"

[project.urls]
Homepage = "https://github.com/pyfunc/taskfile"
Repository = "https://github.com/pyfunc/taskfile"
Issues = "https://github.com/pyfunc/taskfile/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"taskfile.scaffold" = ["templates/*.yml"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/qwen/qwen3-coder-next"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.costs]
# AI Cost tracking configuration
badge = true
update_readme = true
readme_path = "README.md"
default_model = "openrouter/qwen/qwen3-coder-next"
analysis_mode = "byok"
full_history = true
max_commits = 500

# Cost thresholds for badge colors (USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }
